requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.12.3
cachetools==5.2.0
certifi==2020.12.5
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.0
cssselect2==0.7.0
distlib==0.3.8
Django==4.1.3
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-htmlmin==0.11.0
django-minify-html==1.7.1
django-storages==1.13.1
django-xhtml2pdf==0.0.4
docopt==0.6.2
filelock==3.13.4
fpdf==1.7.2
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
html5lib==1.1
httplib2==0.20.4
idna==2.10
lxml==5.1.0
minify_html==0.15.0
num2words==0.5.10
numpy==1.26.4
oauthlib==3.2.1
opencv-python==4.9.0.80
oscrypto==1.3.0
packaging==24.0
Pillow==9.2.0
platformdirs==4.2.0
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyparsing==3.0.9
pypdf==4.0.0
pypng==0.20220715.0
pytesseract==0.3.10
python-bidi==0.4.2
pytz==2021.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.4
reportlab-qrcode==0.1.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.1
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.3
virtualenv==20.25.3
webencodings==0.5.1
xhtml2pdf==0.2.14
xlwt==1.3.0
